dns: fix validation copypasta error.
Names that do not end in [username].hamwan.net were causing an error instead
of returning to appropraite validation message. It was only checking is the
reverse was a valid PTR, so things like A records, not being proper reverses,
would cause an error when passed to the "ip in subnet" test. Now it skips the
test when it's not an IP address, continuing straight to the validation
message.