Add management command discoverclients to add new client radios to the portal.
Run with

    ./manage.py discoverclients

This command uses SNMP queries to sector routers to discover clients. MACs
discovered via SNMP are compared to portal entries. If the MAC is not found,
it creates a Host record and IPAddress record for the
previously-undocumented client.

There are some caveats and missing features. Known issues are documented in
the code.