Support more OSes, and improve Ansible group names
OSes now include "None", which will translate to NULL in SQL.

Ansible group names are now type-prefixed to minimize chances of
collision.  Previously, if someone added a host named "linux" there
would be a collision with the "linux" OS group.

Entities with NULL OSes are still easily selectable as just "os_".

Entities with "os_other" should be considered as a trigger to expand the
OS_TYPES list in models.py.